refactor: move login modal logic to custom hook
